Try to fix fonts, unsure
